requirements.txt
# torch>=2.0
# torchvision
# numpy==1.26.4
# opencv-python-headless
# pillow
# transformers>=4.36.0
# diffusers==0.25.0
# controlnet_aux==0.0.7
# streamlit
# huggingface_hub==0.19.4

streamlit
torch
transformers
diffusers
opencv-python-headless
pillow
huggingface_hub